refactor: clarify site web boundaries
