+++ crates/site/server/src/router_tests.rs
@@ -6,7 +6,7 @@ use std::{
     },
     time::SystemTime,
 };
-use web::topcoat_pages;
+use web::assets;
 
 use async_trait::async_trait;
 use domain::{
@@ -57,19 +57,19 @@ fn test_asset_config() -> AssetConfig {
                     content_type: "text/javascript".to_string(),
                 },
                 ManifestEntry {
-                    id: topcoat_pages::CLIENT_NAVIGATION_SCRIPT.id(),
-                    file: "topcoat-navigation-test.js".to_string(),
+                    id: assets::NAVIGATION_SCRIPT.id(),
+                    file: "navigation-test.js".to_string(),
                     hash: "test".to_string(),
                     content_type: "text/javascript".to_string(),
                 },
                 ManifestEntry {
-                    id: topcoat_pages::STYLESHEET.id(),
+                    id: assets::STYLESHEET.id(),
                     file: "tailwind-test.css".to_string(),
                     hash: "test".to_string(),
                     content_type: "text/css".to_string(),
                 },
                 ManifestEntry {
-                    id: topcoat_pages::FAVICON.id(),
+                    id: assets::FAVICON.id(),
                     file: "favicon-test.ico".to_string(),
                     hash: "test".to_string(),
                     content_type: "image/x-icon".to_string(),
@@ -340,7 +340,7 @@ async fn home_renders_the_published_summary_as_html() {
 }
 
 #[tokio::test]
-async fn home_shell_exposes_topcoat_mobile_navigation_contract() {
+async fn home_shell_exposes_mobile_navigation_contract() {
     let router = create_router(fixture_reader(), false);
     let response = response(
         &router,
@@ -354,9 +354,11 @@ async fn home_shell_exposes_topcoat_mobile_navigation_contract() {
             .body
             .contains("<script type=\"module\" src=\"/_topcoat/assets/topcoat-test.js\"></script>")
     );
-    assert!(response.body.contains(
-        "<script type=\"module\" src=\"/_topcoat/assets/topcoat-navigation-test.js\"></script>"
-    ));
+    assert!(
+        response.body.contains(
+            "<script type=\"module\" src=\"/_topcoat/assets/navigation-test.js\"></script>"
+        )
+    );
     assert!(response.body.contains("aria-controls=\"site-header-nav\""));
     assert!(response.body.contains("aria-expanded=\"false\""));
     assert!(
@@ -996,7 +998,7 @@ async fn about_renders_the_published_page_as_html() {
 }
 
 #[tokio::test]
-async fn about_shell_initializes_generated_content_renderers() {
+async fn about_shell_initializes_content_enhancements() {
     let router = create_router(fixture_reader(), false);
     let response = response(
         &router,
