+++ crates/site/web/src/lib.rs
@@ -1,47 +1,20 @@
 extern crate self as web;
 
-pub mod format;
-pub mod generated_content;
-pub mod topcoat_pages;
+mod article_card;
+pub mod assets;
+mod content_enhancement;
+mod format;
+mod page_loader;
+pub mod pages;
+mod shell;
 
-use std::sync::Arc;
+pub use page_loader::{DynPageLoader, PageLoadResult, PageLoader, PageLoaderContext};
 
-use async_trait::async_trait;
-use domain::{
-    ArticlePageDocument, Category, CategoryPageDocument, HomePageDocument, PageKey, Slug,
-    StaticPageDocument,
-};
-
-pub const SITE_NAME: &str = "ぶくせんの探窟メモ";
-pub const SITE_ORIGIN: &str = "https://www.okawak.net";
+pub(crate) const SITE_NAME: &str = "ぶくせんの探窟メモ";
+pub(crate) const SITE_ORIGIN: &str = "https://www.okawak.net";
 const SITE_ORIGIN_ENV: &str = "OKAWAK_BLOG_SITE_ORIGIN";
 
-pub type PageLoadResult<T> = Result<T, String>;
-
-#[async_trait]
-pub trait PageLoader: Send + Sync {
-    async fn load_home(&self) -> PageLoadResult<HomePageDocument>;
-
-    async fn load_article(
-        &self,
-        category: &Category,
-        slug: &Slug,
-    ) -> PageLoadResult<Option<ArticlePageDocument>>;
-
-    async fn load_category(
-        &self,
-        category: &Category,
-    ) -> PageLoadResult<Option<CategoryPageDocument>>;
-
-    async fn load_static_page(&self, page: &PageKey) -> PageLoadResult<Option<StaticPageDocument>>;
-}
-
-pub type DynPageLoader = Arc<dyn PageLoader>;
-
-#[derive(Clone)]
-pub struct PageLoaderContext(pub DynPageLoader);
-
-pub fn build_site_url(path: &str) -> String {
+pub(crate) fn build_site_url(path: &str) -> String {
     join_site_url(&resolved_site_origin(), path)
 }
